format prompting cleaned
